Rename sveltekit adapter to `paraglide-sveltekit`
