+++ inlang/guides/build-a-global-svelte-app/build-a-global-svelte-app.md
@@ -158,7 +158,7 @@ Paraglide adapters are framework-specific packages that make it easier to use Pa
 Install the SvelteKit adapter:
 
 ```cmd
-npm i @inlang/paraglide-js-adapter-sveltekit
+npm i @inlang/paraglide-sveltekit
 ```
 
 
@@ -168,7 +168,7 @@ The SvelteKit adapter provides a Vite plugin that automatically compiles your tr
 
 ```ts
 import { sveltekit } from "@sveltejs/kit/vite"
-import { paraglide } from "@inlang/paraglide-js-adapter-sveltekit/vite"
+import { paraglide } from "@inlang/paraglide-sveltekit/vite"
 import { defineConfig } from "vite"
 
 export default defineConfig({
@@ -188,7 +188,7 @@ To initialize the Adapter, we need to call `createI18n` and pass it the paraglid
 
 ```ts
 // ./src/lib/i18n.ts
-import { createI18n } from "@inlang/paraglide-js-adapter-sveltekit"
+import { createI18n } from "@inlang/paraglide-sveltekit"
 import * as runtime from "$paraglide/runtime"
 
 export const i18n = createI18n(runtime)
@@ -201,7 +201,7 @@ To provide the language to your app, add the `ParaglideJS` component to your lay
 ```svelte
 <!-- src/routes/+layout.svelte -->
 <script>
-	import { ParaglideJS } from '@inlang/paraglide-js-adapter-sveltekit'
+	import { ParaglideJS } from '@inlang/paraglide-sveltekit'
 	import { i18n } from '$lib/i18n'
 </script>
 
@@ -299,5 +299,3 @@ Try it on [StackBlitz](https://stackblitz.com/~/github.com/lorissigrist/paraglid
 If you have any questions, feel free to ask them in our [Discord](https://discord.gg/CNPfhWpcAa) or open a discussion on [GitHub](https://github.com/opral/monorepo/discussions).
 
 You can reference an example project on [GitHub](https://github.com/opral/monorepo/tree/main/inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example).
-
-This guide is based on `paraglide-js 1.1.0`, `paraglide-js-adapter-sveltekit 0.1.0`, `plugin-message-format 2.0.0`, `m-function-matcher 0.5.0.` and `paraglide-js-adapter-vite 1.0.0-prerelease.2`.
