+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/app/createI18n.server.ts
@@ -18,7 +18,7 @@ import { PrefixStrategy } from "./routing/prefixStrategy"
  * ```ts
  * // src/lib/i18n.js:
  * import * as runtime from "../paraglide/runtime.js"
- * import { createI18n } from "@inlang/paraglide-js-adapter-sveltekit"
+ * import { createI18n } from "@inlang/paraglide-sveltekit"
  *
  * export const i18n = createI18n({ ...options })
  * ```
