+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @inlang/paraglide-js-adapter-sveltekit
+# @inlang/paraglide-sveltekit
 
 ## 0.6.7
