+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/runtime/adapter.ts
@@ -146,7 +146,7 @@ export type I18nConfig<T extends string> = {
  * ```ts
  * // src/lib/i18n.js
  * import * as runtime from "../paraglide/runtime.js"
- * import { createI18n } from "@inlang/paraglide-js-adapter-sveltekit"
+ * import { createI18n } from "@inlang/paraglide-sveltekit"
  *
  * export const i18n = createI18n(runtime, { ...options })
  * ```
